requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
Django==3.0.3
fake-useragent==0.1.11
idna==2.9
lxml==4.5.0
mysqlclient==1.4.6
numpy==1.18.1
pandas==1.0.1
parse==1.14.0
pyee==7.0.1
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-html==0.10.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tqdm==4.43.0
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
yahoo-fin==0.8.4
djangorestframework==3.11.0
django-filter==2.2.0
